+++ app.js
@@ -89,6 +89,7 @@ socketManager.authMiddleware = function (socket, next) {
 }
 
 let gw // the gateway instance
+const plugins = []
 
 // flag used to prevent multiple restarts while one is already in progress
 let restarting = false
@@ -249,7 +250,39 @@ function startGateway (settings) {
 
   gw.start()
 
+  const pluginsConfig = settings.gateway ? settings.gateway.plugins : null
+
+  // load custom plugins
+  if (pluginsConfig && Array.isArray(pluginsConfig)) {
+    for (const plugin of pluginsConfig) {
+      try {
+        const pluginName = path.basename(plugin)
+        const instance = require(plugin)({
+          zwave,
+          mqtt,
+          app,
+          logger: loggers.module(pluginName)
+        })
+        instance.name = pluginName
+        plugins.push(instance)
+        logger.info(`Successfully loaded plugin ${instance.name}`)
+      } catch (error) {
+        logger.error(`Error while loading ${plugin} plugin`, error)
+      }
+    }
+  }
+
   restarting = false
+}
+
+async function destroyPlugins () {
+  while (plugins.length > 0) {
+    const instance = plugins.pop()
+    if (instance && typeof instance.destroy === 'function') {
+      logger.info('Closing plugin ' + instance.name)
+      await instance.destroy()
+    }
+  }
 }
 
 function setupInterceptor () {
@@ -330,7 +363,7 @@ function setupSocket (server) {
   socketManager.on(inboundEvents.init, function (socket) {
     if (gw.zwave) {
       socket.emit(socketEvents.init, {
-        nodes: gw.zwave.nodes,
+        nodes: gw.zwave.getNodes(),
         info: gw.zwave.getInfo(),
         error: gw.zwave.error,
         cntStatus: gw.zwave.cntStatus
@@ -346,31 +379,82 @@ function setupSocket (server) {
     }
   })
 
+  socketManager.on(inboundEvents.mqtt, async function (socket, data) {
+    logger.info(`Mqtt api call: ${data.api}`)
+
+    let res, err
+
+    try {
+      switch (data.api) {
+        case 'updateNodeTopics':
+          res = gw.updateNodeTopics(data.args[0])
+          break
+        case 'removeNodeRetained':
+          res = gw.removeNodeRetained(data.args[0])
+          break
+        default:
+          err = `Unknown MQTT api ${data.apiName}`
+      }
+    } catch (error) {
+      logger.error('Error while calling MQTT api', error)
+      err = error.message
+    }
+
+    const result = {
+      success: !err,
+      message: err || 'Success MQTT api call',
+      result: res
+    }
+    result.api = data.api
+
+    socket.emit(socketEvents.api, result)
+  })
+
   socketManager.on(inboundEvents.hass, async function (socket, data) {
     logger.info(`Hass api call: ${data.apiName}`)
-    switch (data.apiName) {
-      case 'delete':
-        gw.publishDiscovery(data.device, data.nodeId, true, true)
-        break
-      case 'discover':
-        gw.publishDiscovery(data.device, data.nodeId, false, true)
-        break
-      case 'rediscoverNode':
-        gw.rediscoverNode(data.nodeId)
-        break
-      case 'disableDiscovery':
-        gw.disableDiscovery(data.nodeId)
-        break
-      case 'update':
-        gw.zwave.updateDevice(data.device, data.nodeId)
-        break
-      case 'add':
-        gw.zwave.addDevice(data.device, data.nodeId)
-        break
-      case 'store':
-        await gw.zwave.storeDevices(data.devices, data.nodeId, data.remove)
-        break
+
+    let res, err
+    try {
+      switch (data.apiName) {
+        case 'delete':
+          res = gw.publishDiscovery(data.device, data.nodeId, true, true)
+          break
+        case 'discover':
+          res = gw.publishDiscovery(data.device, data.nodeId, false, true)
+          break
+        case 'rediscoverNode':
+          res = gw.rediscoverNode(data.nodeId)
+          break
+        case 'disableDiscovery':
+          res = gw.disableDiscovery(data.nodeId)
+          break
+        case 'update':
+          res = gw.zwave.updateDevice(data.device, data.nodeId)
+          break
+        case 'add':
+          res = gw.zwave.addDevice(data.device, data.nodeId)
+          break
+        case 'store':
+          res = await gw.zwave.storeDevices(
+            data.devices,
+            data.nodeId,
+            data.remove
+          )
+          break
+      }
+    } catch (error) {
+      logger.error('Error while calling HASS api', error)
+      err = error.message
     }
+
+    const result = {
+      success: !err,
+      message: err || 'Success HASS api call',
+      result: res
+    }
+    result.api = data.apiName
+
+    socket.emit(socketEvents.api, result)
   })
 }
 
@@ -618,26 +702,37 @@ app.post('/api/importConfig', apisLimiter, isAuthenticated, async function (
   req,
   res
 ) {
-  const config = req.body.data
+  let config = req.body.data
   try {
     if (!gw.zwave) throw Error('Zwave client not inited')
 
-    if (!Array.isArray(config)) throw Error('Configuration not valid')
-    else {
+    // try convert to node object
+    if (Array.isArray(config)) {
+      const parsed = {}
+
       for (let i = 0; i < config.length; i++) {
-        const e = config[i]
-        if (
-          e &&
-          (!utils.hasProperty(e, 'name') || !utils.hasProperty(e, 'loc'))
-        ) {
-          continue
-        } else if (e) {
-          await gw.zwave.callApi('_setNodeName', i, e.name || '')
-          await gw.zwave.callApi('_setNodeLocation', i, e.loc || '')
-          if (e.hassDevices) {
-            await gw.zwave.storeDevices(e.hassDevices, i, false)
-          }
+        if (config[i]) {
+          parsed[i] = config[i]
         }
+      }
+
+      config = parsed
+    }
+
+    for (const nodeId in config) {
+      const node = config[nodeId]
+      if (!node || typeof node !== 'object') continue
+
+      if (utils.hasProperty(node, 'name')) {
+        await gw.zwave.callApi('setNodeName', nodeId, node.name || '')
+      }
+
+      if (utils.hasProperty(node, 'loc')) {
+        await gw.zwave.callApi('setNodeLocation', nodeId, node.loc || '')
+      }
+
+      if (node.hassDevices) {
+        await gw.zwave.storeDevices(node.hassDevices, nodeId, false)
       }
     }
 
@@ -806,6 +901,7 @@ app.post('/api/settings', apisLimiter, isAuthenticated, async function (
     restarting = true
     await jsonStore.put(store.settings, settings)
     await gw.close()
+    await destroyPlugins()
     // reload loggers settings
     setupLogging(settings)
     // restart clients and gateway
@@ -849,6 +945,7 @@ async function gracefuShutdown () {
   logger.warn('Shutdown detected: closing clients...')
   try {
     await gw.close()
+    await destroyPlugins()
   } catch (error) {
     logger.error('Error while closing clients', error)
   }